+++ R/get_adjacencies.R
@@ -5,7 +5,7 @@
 #' @param landscape Raster* Layer, Stack, Brick or a list of rasterLayers.
 #' @param neighbourhood The number of directions in which cell adjacencies are
 #' considered as neighbours: 4 (rook's case), 8 (queen's case) or a binary matrix
-#' where 1 define the neighbourhood and 0 the focal cell. The default is 4.
+#'  where the ones define the neighbourhood. The default is 4.
 #' @param what Which adjacencies to calculate: "full" for a full adjacency matrix,
 #' "like" for the diagonal, "unlike" for the off diagonal part of the matrix and
 #'  "triangle" for a triangular matrix counting adjacencies only once.
@@ -27,7 +27,17 @@
 #'
 #' @examples
 #' # calculate full adjacency matrix
-#' get_adjacencies(landscape, neighbourhood = 4)
+#' get_adjacencies(landscape, 4)
+#'
+#' # count diagonal neighbour adjacencies
+#' diagonal_matrix <- matrix(c(1,  NA,  1,
+#'                             NA,  0, NA,
+#'                             1,  NA,  1), 3, 3, byrow = TRUE)
+#' get_adjacencies(landscape, diagonal_matrix)
+#'
+#' # equivalent with the raster package:
+#' adjacencies <- raster::adjacent(landscape, 1:raster::ncell(landscape), 4, pairs=TRUE)
+#' table(landscape[adjacencies[,1]], landscape[adjacencies[,2]])
 #'
 #' @aliases get_adjacencies
 #' @rdname get_adjacencies
@@ -38,12 +48,12 @@ get_adjacencies <- function(landscape,
                             what = "full",
                             upper = FALSE){
 
-    if(neighbourhood != 4 && neighbourhood != 8 && !is.matrix(neighbourhood)){
+    if(!identical(neighbourhood, 4) && !identical(neighbourhood, 8) && !is.matrix(neighbourhood)){
      stop("neighbourhood must be either 4, 8 or a binary matrix where the ones define the neighbourhood.", call. = FALSE)
     }
 
-    adjacencies <- rcpp_get_coocurrence_matrix(x = raster::as.matrix(landscape),
-                                               directions = as.matrix(neighbourhood))
+    adjacencies <- rcpp_get_coocurrence_matrix(raster::as.matrix(landscape),
+                                as.matrix(neighbourhood))
 
     if (!isTRUE(upper)) {
         if (what == "like") {