+++ R/list_lsm.R
@@ -10,10 +10,10 @@
 #' It is also possible to specify functions as a vector of strings, e.g. `what = c("lsm_c_ca", "lsm_l_ta")`.
 #' @param simplify If true, function names are returned as vector.
 #' @param verbose Print warning messages
-
 #'
-#' @details List all available landscape metrics depending on the provided filter arguments.
-#' If an argument is not provided, automatically all possibilities are selected.
+#' @details List all available landscape metrics depending on the provided filter
+#' arguments. If an argument is not provided, automatically all possibilities are
+#' selected. Therefore, to get **all** available metrics, use simply `list_lsm()`.
 #' Furthermore, it is possible to only get a vector with all function names instead of the full tibble.
 #'
 #' @return tibble