+++ R/lsm_p_circle.R
@@ -164,7 +164,8 @@ lsm_p_circle_calc <- function(landscape, directions,
         landscape_labeled <- t(landscape_labeled)
 
         # get coordinates of current class
-        points <- points[which(!is.na(class_boundaries)), ]
+        points <- matrix(points[which(!is.na(class_boundaries)), ],
+                         ncol = 3)
 
         # use patch id instead of class id
         points[, 3] <- landscape_labeled[!is.na(class_boundaries)]
@@ -201,87 +202,3 @@ lsm_p_circle_calc <- function(landscape, directions,
         value = as.double(circle_patch$value)
     )
 }
-
-# lsm_p_circle_calc_old <- function(landscape, directions,
-#                                   extent = NULL, resolution = NULL, crs = NULL) {
-#
-#     # use raster instead of landscape
-#     if(class(landscape) == "matrix") {
-#         landscape <- matrix_to_raster(landscape,
-#                                       extent = extent,
-#                                       resolution = resolution,
-#                                       crs =crs)
-#     }
-#
-#     # get resolution of landscape
-#     resolution <- raster::res(landscape)
-#     resolution_x <- resolution[[1]]
-#     resolution_y <- resolution[[2]]
-#
-#     # get patch area
-#     area_patch <- lsm_p_area_calc(landscape, directions = directions)
-#
-#     # patches with only 1 cell
-#     one_cell <- which(area_patch$value == prod(resolution) / 10000 )
-#
-#     # get unique classes
-#     classes <- get_unique_values(landscape)[[1]]
-#
-#     circle_patch <- lapply(classes, function(patches_class) {
-#
-#         # get connected patches
-#         landscape_labeled <- get_patches(landscape,
-#                                          class = patches_class,
-#                                          directions = directions)[[1]]
-#
-#         # only boundary cells need to be considered for circle diameter
-#         landscape_boundaries <- get_boundaries(landscape_labeled,
-#                                                directions = 4,
-#                                                as_NA = TRUE)
-#
-#         # convert to points
-#         points_class_labeled <- data.frame(raster::rasterToPoints(landscape_labeled))
-#         points_class_boundaries <- data.frame(raster::rasterToPoints(landscape_boundaries))
-#
-#         # keep only points that are boundary (but with original patch id)
-#         points_class <- dplyr::semi_join(x = points_class_labeled,
-#                                          y = points_class_boundaries,
-#                                          by = c("x","y"))
-#
-#         # get circle radius around patch
-#         circle <- rcpp_get_circle(as.matrix(points_class),
-#                                   resolution_x = resolution_x,
-#                                   resolution_y = resolution_y)
-#         # calculate circle area
-#         circle[, 2] <- pi * ((circle[, 2]  /2) ^ 2)
-#
-#         # sort according to patch id
-#         circle <- matrix(circle[order(circle[,1]),], ncol = 2)
-#
-#         tibble::tibble(class = patches_class,
-#                        value = circle[,2])
-#     })
-#
-#     # calculate circle metric
-#     circle_patch <- dplyr::bind_rows(circle_patch)
-#
-#     # calculate circle metric
-#     circle_patch <- dplyr::mutate(circle_patch,
-#                                   value = 1 - ((area_patch$value * 10000) / value))
-#
-#     # set all one-cell patches to 0
-#     circle_patch$value[one_cell] <- 0
-#
-#     tibble::tibble(
-#         level = "patch",
-#         class = as.integer(circle_patch$class),
-#         id = as.integer(seq_len(nrow(circle_patch))),
-#         metric = "circle",
-#         value = as.double(circle_patch$value)
-#     )
-# }
-
-# bench::mark(
-#     new <- lsm_p_circle_calc(landscape, directions = 8),
-#     old <- lsm_p_circle_calc_old(landscape, directions = 8),
-#     iterations = 25)